+++ src/Symfony/Component/HttpKernel/Tests/EventListener/LocaleListenerTest.php
@@ -12,10 +12,12 @@
 namespace Symfony\Component\HttpKernel\Tests\EventListener;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\EventListener\LocaleListener;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
+use Symfony\Component\HttpKernel\KernelEvents;
 
 class LocaleListenerTest extends TestCase
 {
@@ -26,12 +28,28 @@ class LocaleListenerTest extends TestCase
         $this->requestStack = $this->getMockBuilder('Symfony\Component\HttpFoundation\RequestStack')->disableOriginalConstructor()->getMock();
     }
 
-    public function testDefaultLocaleWithoutSession()
+    public function testIsAnEventSubscriber()
+    {
+        $this->assertInstanceOf(EventSubscriberInterface::class, new LocaleListener($this->requestStack));
+    }
+
+    public function testRegisteredEvent()
+    {
+        $this->assertEquals(
+            array(
+                KernelEvents::REQUEST => array(array('setDefaultLocale', 100), array('onKernelRequest', 16)),
+                KernelEvents::FINISH_REQUEST => array(array('onKernelFinishRequest', 0)),
+            ),
+            LocaleListener::getSubscribedEvents()
+        );
+    }
+
+    public function testDefaultLocale()
     {
         $listener = new LocaleListener($this->requestStack, 'fr');
         $event = $this->getEvent($request = Request::create('/'));
 
-        $listener->onKernelRequest($event);
+        $listener->setDefaultLocale($event);
         $this->assertEquals('fr', $request->getLocale());
     }